+++ Assets/Scripts/GameStructure/Narrative.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: caca71c67e5c2f143926ea0a986439c0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
