+++ Assets/Scripts/GameStructure/Narrative/NarrativeLayout.cs.meta
