+++ Assets/Scripts/GameStructure/Narrative/NarrativePanel.cs.meta
