--- Assets/Scripts/GameStructure/NarrativePanel.cs
@@ -1,12 +1,0 @@
-using System;
-using UnityEngine;
-
-[CreateAssetMenu]
-[Serializable]
-public class NarrativePanel : ScriptableObject
-{
-    // Contains the information about a specific panel for displaying story information on the screen
-    public string simplePanelName;
-    public bool forceFullScreen; // Use this to force it to be the only element on screen regardless of NarrativeSO settings
-    public Sprite panelImage; // if this is null we use the default layout for panels
-}
