+++ Assets/Scripts/Managers/NarrativeDataManager.cs
@@ -1,3 +1,4 @@
+using GameStructure.Narrative;
 using UnityEngine;
 using Utils;
 
@@ -5,11 +6,11 @@ namespace Managers
 {
     public class NarrativeDataManager : Singleton<NarrativeDataManager>
     {
-        [SerializeField] private NarrativeSO[] allNarrativeData;
+        [SerializeField] private NarrativeSo[] allNarrativeData;
         
-        public NarrativeSO GetNarrativeData(int actNumber, NarrativeSO.NarrativeType narrativeType)
+        public NarrativeSo GetNarrativeData(int actNumber, NarrativeSo.NarrativeType narrativeType)
         {
-            foreach (NarrativeSO narrativeSo in allNarrativeData)
+            foreach (NarrativeSo narrativeSo in allNarrativeData)
             {
                 if (narrativeSo.actNumber == actNumber && narrativeSo.narrativeType == narrativeType)
                 {
